+++ src/3rd_party/sol_config.hpp
@@ -27,27 +27,7 @@
 #ifndef SOL_SINGLE_CONFIG_HPP
 #define SOL_SINGLE_CONFIG_HPP
 
-// beginning of sol/config.hpp
-
-/* Base, empty configuration file!
-
-     To override, place a file in your include paths of the form:
-
-. (your include path here)
-| sol (directory, or equivalent)
-  | config.hpp (your config.hpp file)
-
-     So that when sol2 includes the file
-
-#include <sol/config.hpp>
-
-     it gives you the configuration values you desire. Configuration values can be
-seen in the safety.rst of the doc/src, or at
-https://sol2.readthedocs.io/en/latest/safety.html ! You can also pass them through
-the build system, or the command line options of your compiler.
-
-*/
-
-// end of sol/config.hpp
+#define SOL_ALL_SAFETIES_ON 1
+#define SOL_SAFE_FUNCTION 1
 
 #endif // SOL_SINGLE_CONFIG_HPP
