+++ src/episode/episodeclass.hpp
@@ -62,9 +62,6 @@ struct PK2EPISODESCORES10 {
 	
 };
 
-namespace sol{
-	class state;
-}
 
 class EpisodeClass {
 
@@ -111,9 +108,6 @@ class EpisodeClass {
 		void Update_NextLevel();
 
 		SfxHandler sfx;
-
-		sol::state * lua = nullptr;
-
 	private:
 		int  Open_Scores();
